Moves hash_to_point into the ring_signature module
