+++ transaction/core/src/ring_signature/mod.rs
@@ -4,11 +4,15 @@
 #![macro_use]
 extern crate alloc;
 
+use crate::domain_separators::HASH_TO_POINT_DOMAIN_TAG;
+use blake2::{Blake2b, Digest};
 use bulletproofs::{BulletproofGens, PedersenGens};
+use curve25519_dalek::ristretto::RistrettoPoint;
 pub use curve25519_dalek::scalar::Scalar;
 pub use curve_scalar::*;
 pub use error::Error;
 pub use key_image::*;
+use mc_crypto_keys::RistrettoPublic;
 pub use mlsag::*;
 pub use rct_bulletproofs::*;
 
@@ -27,4 +31,12 @@ lazy_static! {
     /// be at least 2 * MAX_INPUTS + MAX_OUTPUTS, which allows for inputs, pseudo outputs, and outputs.
     pub static ref BP_GENERATORS: BulletproofGens =
         BulletproofGens::new(64, 64);
+}
+
+/// Applies a hash function and returns a RistrettoPoint.
+pub fn hash_to_point(ristretto_public: &RistrettoPublic) -> RistrettoPoint {
+    let mut hasher = Blake2b::new();
+    hasher.update(&HASH_TO_POINT_DOMAIN_TAG);
+    hasher.update(&ristretto_public.to_bytes());
+    RistrettoPoint::from_hash(hasher)
 }
